Extracted template from most-commented-films component
