+++ src/components/most-commented-films/template.js
@@ -1,0 +1,13 @@
+/**
+ * Creates top rated films template
+ * @return {String} template
+ */
+const template = () => {
+  return `<section class="films-list--extra films-list--most-commented">
+            <h2 class="films-list__title">Most commented</h2>
+            <div class="films-list__container">
+            </div>
+          </section>`;
+};
+
+export {template};
